requirements.txt
python-jobspy
python-telegram-bot
sqlmodel
groq
instructor
python-dotenv
pandas